Rename PerfDataReader to PerfByteReader
PerfDataReader sounds like it's a parser for perf.data files, but it
isn't. It's just the byteswapper helper used by this Perf helper
library. Rename it to PerfByteReader in hopes that's a little bit less
confusing.

Also add a line to the CMakeLists.txt for the recently-added
PerfEventInfo.h header (should have been added in the previous commit).