Prefer TraceFS over DebugFS (#33)
* Prefer TraceFS over DebugFS

When looking for the tracing filesystem, we scan the mounted filesystems
for either "tracefs" or "debugfs". We can use either one.

Old behavior is to take the first matching mount point, which might be
either tracefs or debugfs. This can be confusing if the user has both
debugfs and tracefs mounted - which one gets used? (Depends on which one
comes first in the list of mounted filesystems.)

New behavior is to prefer tracefs over debugfs. That way, if both
tracefs and debugfs are mounted, we consistently use tracefs, making the
behavior more predictable.