Include clock information in event metadata
The existing library tracks event timestamp as a nanosecond offset from
some unknown reference time. This allows determining event order and the
time between events, but it does not help when determining the date/time
an event occurred. For that, we need to know the wall-clock time
corresponding to that reference time.

This PR adds support for tracking the "clock information" for a session,
which includes the clock type (e.g. MONOTONIC or REALTIME) and the
wall-clock date/time that a `0` timestamp value indicates. This
information may not always be available.

Details:

- Define a type for session metadata. Include clock information in this
  type.
- Add a pointer to the session metadata to the event metadata.
- When parsing a `perf.data` file that contains clock information,
  record it in the session metadata. (`perf.data` file will contain
clock information if the `-k` option was used, `perf record -k mono ...`).
- When running a live session, record the clock information in the
  session metadata.
- When formatting an event, if clock information is available, format as
  `"yyyy-mm-ddThh:mm:ss.nnnnnnnnnZ"`. If clock information is not
  available, fall back to the existing `ss.nnnnnnnnn` format.
- Update version to 1.3 for the affected libraries. Unaffected libraries
  remain at 1.2.