Probe for debugfs is still needed
While `debugfs` will auto-mount `tracefs`, it does so lazily - `tracefs`
won't show up in `/proc/mounts` until somebody tries to use it. This
means we can't depend on always finding `tracefs` in `/proc/mounts`, so
we do need to be able to fall-back to the actual `debugfs` entry.

In addition, the direct probe in TracepointPath.cpp can have a false
positive - the `/sys/kernel/tracing` directory can exist without
containing a valid mounted `tracefs`. We need to look for something
inside the directory if we want an accurate probe, so look for the
`events` subdirectory.