TracepointSession - provide tracepoint stats (#40)
Requirement: Need to be able to identify and disable overly-noisy events
from the session.

Task: Provide a way to obtain statistics for each tracepoint enabled
into a session, e.g. `session.TracepointInfo(tracepointName).Count()`.

Technical: The `tracepoint` system does keep a count on a
per-tracepoint per-CPU basis. The count is obtained by doing a `read`
from the file handle corresponding to the tracepoint enablement.

Overview:

- Create a TracepointInfo class that exposes information about a
  single tracepoint that is part of the session. This class allows
  reading the enable/disable state, the metadata, and the count (the
  number of times a particular tracepoint has been written to the session,
  totaled across all CPUs).
- Support enumerating all of the TracepointInfo objects in the session.
- Support accessing a specific TracepointInfo by ID or by name.

Note that this kind of count statistic is unique to Linux tracepoints
and is not available with ETW.