TracepointSession: SavePerfDataFile method
Add support for saving a `perf.data` format file from a session.

- Implementation detail: Change session enumeration internals to provide
  the enumeration handlers with access to the event's header.
- Don't have separate enumeration handlers for sample and non-sample
  events. Just have one handler that makes its own decisions.
- Add a sample that demonstrates saving events from a session.
- Include a pointer to the raw event in the data we provide during event
  enumeration.

Still TODO: This provides just the data needed for decoding traces. It
does not include any of the optional system-information headers that
would be collected by the `perf` tool (similar to ETW rundown info).
Most of the headers are probably unnecessary for basic tracepoint
analysis, but there are a few that might be worth adding.