TracepointSession - allow per-CPU buffer sizes (#45)
* TracepointSession - allow per-CPU buffer sizes

Issue: we want to minimize memory usage. We know that some of the trace
producers will be bound to specific CPUs. If there are no producers for
a particular CPU, it is wasteful to allocate a buffer to collect data
from that CPU.

Old: TracepointSession allocates one buffer per CPU. Each buffer is the
same size.

New: TracepointSession allows each buffer to be a different size. Size
can even be 0 if we don't need to accept events from that CPU.

Details:

- Existing code will continue to get the same buffer size for all CPUs.
- Add new index parameter to the BufferSize(index) getter so you can get
  a per-CPU size. The parameter defaults to 0 for the common case when
  all buffers have the same size.
- You can use a new constructor for `TracepointSessionOptions` to
  provide a list of buffer sizes rather than a single buffer size.
- If the list is too short, or if an element of the list is 0, the
  buffer size for the corresponding CPU will be 0 (disable collection).
- Rest of the changes are just to use a per-buffer size and to deal with
  the possibility that a buffer size may be 0 (no tracing on that buffer).

In addition, changes trace session to use `PERF_ATTR_SIZE_VER3` as the
structure size for `perf_event_attr` when calling `perf_event_open` to
minimize the possibility of compatibility issues on older kernels.

* Update BufferInfo constructor with new fields