TracepointSession - don't uncessarily track empty buffers
A few of the data structures and many of the loops in TracepointSession
are sized by m_bufferCount, e.g. m_buffers points at an array of
m_bufferCount items.

In some cases, some of the slots in m_buffers may be empty. If the empty
slots are at the end of the array, there is no point in tracking them.
In this case, we should reduce the size of the array so we allocate less
memory and don't have to loop through empty buffers.