decode-perf - show event name even when no EVENT_DESC (#68)
Summary:

- Formatting code should look for provider/event names in format metadata
  if they aren't available in the EVENT_DESC metadata.
- Correct comments indicating that FlushToWriter will write a
  FinishedRound event. I was thinking about doing that, and wrote the
  comment accordingly, but later decided against it and forgot to update
  the comment.
- Add missing `#include <unistd.h>`. (Whether or not it's actually
  needed depends on the SDK you're building against, which is how it got
  missed before.)
- Update comments to point out that event name might be "" if EVENT_DESC
  header is not present, and to note that the caller might want to use
  format metadata as a fallback.

Background:

For non-eventheader events, the event name comes from the EVENT_DESC
header in the perf.data file. That header is optional. When it's
missing, the previous version of decode-perf didn't get event names for
non-eventheader events.

However, most tracepoints have associated format information, and that
has the system name and the tracepoint name. For tracepoints, the event
name is tracepoint name + ':' + system name, so that means we can
reconstruct the event name from format info  even when EVENT_DESC is missing.

Fix the decode logic to notice when EVENT_DESC is missing and to get the
system name and tracepoint name from format in that case.