Clarify C/C++ language requirements
Some LinuxTracepoints projects do not clearly indicate their minimum
language requirements. This causes problems because different customers
may have a different default language version set for their environment,
and the build will fail if the customer's default is lower than the
unstated requirement.

To fix this, all projects that have specific language version
requirements need to say so. To ensure that it stays fixed, we set our
defaults to be very old. That way, any project that doesn't declare a
requirement is going to get noticed (unless it works on very old
language which is also ok).

- In each root project that builds C (or C++) code, declare the default C
  (or C++) language version to be C90 (or C++98).
- This results in build errors for projects that aren't correctly
  requiring language versions. Add requirements (PUBLIC, INTERFACE, or
  PRIVATE) as appropriate.
- In some cases where a project might have optional headers with higher
  requirements than the lib, add checks in the headers.
- Where appropriate, make code changes to avoid unnecessary requriements.