eventheader - define dynamic and nullable encodings (#71)
EventHeader uses the fixed-length Value128 encoding for IPv6 fields.
This is logically consistent with the other types like GUID and INT32.

ETW uses the variable-length BINARY intype for IPv6 fields. This makes
IPv6 a bit weird because it is nominally fixed-size, but it is
transmitted in a variable-length field. This exposes an interesting
possibility: a BINARY-IPv6 field of length 0 is basically a way to
express "NULL". This "feature" is actually used by customers.

For feature-parity with ETW, EventHeader needs to support this. While
doing so, it probably makes sense to extend this capability to all
types. The basic idea is that you can use a variable-length BINARY
encoding for any type, and a 0 length indicates NULL.

We have existing encoding `string_length16_char8` that can be used for
binary blobs. This is not quite what we need though -- if you use this
encoding with an unrecognized format, the decoder will fall-back to
format `string_utf`, while a real BINARY encoding should fall-back to
`hex_bytes`.

Semantics for variable-length encodings are currently only defined for
blob and string formats. We need to define what happens when you apply a
variable-length encoding to a fixed-length format like UUID or INT.

Changes:

- Define new encoding `binary_length16_char8`. This is exactly the same
  as `string_length16_char8` except that its fallback format is
  `hex_bytes`.
- Define what it means to apply encodings `binary_length16_char8` and
  `string_length16_char8` to a fixed-length format:
  - If the runtime length is 0, the result is a NULL.
  - If the runtime length is supported for the format (e.g. `signed_int`
    supports lengths 1, 2, 4, 8), the format is used.
  - Otherwise, the result is as if the format were `default`.
- Redefine the `ipv4` and `ipv6` formats to both be `ip_address`. If
  applied to a 4-byte field, both should result in an IPv4 format, and
  if applied to a 16-byte field, both should result in an IPv6 format.